[refactor] 유저 한명단 한달에 하나의 예산만 등록가능하게 변경
