+++ src/main/java/dev/book/accountbook/dto/response/BudgetResponse.java
@@ -2,10 +2,12 @@ package dev.book.accountbook.dto.response;
 
 import dev.book.accountbook.entity.Budget;
 
-public record BudgetResponse(int budget, long total) {
+public record BudgetResponse(Long id, int budget, long total) {
     public static BudgetResponse from(Budget budget, int total) {
         return new BudgetResponse(
-                budget.getBudgetLimit(), total
+                budget.getId(),
+                budget.getBudgetLimit(),
+                total
         );
     }
 }
