+++ src/main/java/dev/book/accountbook/repository/querydsl/BudgetRepositoryImpl.java
@@ -1,4 +1,4 @@
-package dev.book.accountbook.repository;
+package dev.book.accountbook.repository.querydsl;
 
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
